+++ lbrynet/daemon/Resources.py
@@ -9,7 +9,7 @@ from twisted.web import server, static, resource
 from twisted.internet import defer, error
 
 from lbrynet import conf
-from lbrynet.lbrynet_daemon.FileStreamer import EncryptedFileStreamer
+from lbrynet.daemon.FileStreamer import EncryptedFileStreamer
 
 
 log = logging.getLogger(__name__)